Clean up stores/mushroom.js
