+++ src/views/HomeView.vue
@@ -1,17 +1,17 @@
 <template>
-  <div id="exploration-grid">
+  <div id="exploration-grid" class="w-100 h-100">
     <ScatterPlot />
     <MushroomCard />
-    <BarChart />
-    <SeasonTable />
+    <ColumnChart />
+    <FloatingRowChart />
   </div>
 </template>
 
 <script setup>
 import ScatterPlot from "@/components/ScatterPlot.vue";
 import MushroomCard from "@/components/card/MushroomCard.vue";
-import SeasonTable from "@/components/SeasonTable.vue";
-import BarChart from "@/components/BarChart.vue";
+import FloatingRowChart from "@/components/FloatingRowChart.vue";
+import ColumnChart from "@/components/ColumnChart.vue";
 </script>
 
 <style lang="sass">
@@ -23,8 +23,6 @@ $padding: 6px
   background: rbg(var(--v-theme-background))
   gap: $padding
   padding: $padding
-  width: 100%
-  height: 100%
 
   #scatter-plot
     order: 1
